templates: split off macro related functions into a separate module
