+++ lib/template/Makefile.am
@@ -2,10 +2,12 @@ templateincludedir			= ${pkgincludedir}/template
 
 templateinclude_HEADERS = \
 	lib/template/templates.h		\
+	lib/template/macros.h			\
 	lib/template/user-function.h
 
 template_sources = \
 	lib/template/templates.c		\
+	lib/template/macros.c			\
 	lib/template/user-function.c
 
 include lib/template/tests/Makefile.am
